feat(admin section): Add vacancies admin section

+++ resources/views/admin/questions/edit.blade.php
@@ -3,17 +3,21 @@
 @section('content')
     <h1>Question {{ $question->id }}</h1>
 
-    <form method="post" action="/admin/questions/edit/{{ $question->id }}" id="edit-question-form" class="row g-3 question-edit-group">
+    <form method="post" action="/admin/questions/{{ $action }}/{{ $question->id }}" id="edit-question-form" class="row g-3 question-edit-group">
         @method('post')
         @csrf
         <div class="col-7">
-            <label for="inputJobVacancy" class="form-label">Job vacancy</label>
-            @error('inputJobVacancy')
-            <span class="invalid-inputJobVacancy" role="alert">
+            <label for="inputVacancy" class="form-label">Job vacancy</label>
+            @error('inputVacancy')
+            <span class="invalid-inputVacancy" role="alert">
                 <strong>{{ $message }}</strong>
             </span>
             @enderror
-            <input type="text" class="form-control" name="inputJobVacancy" id="inputJobVacancy" value="{{ $question->job_vacancy }}" required>
+            <select name="inputVacancy" class="form-select" required>
+                @foreach($vacancies as $vacancy)
+                    <option value="{{ $vacancy->id }}">{{ $vacancy->name }}</option>
+                @endforeach
+            </select>
         </div>
         <div class="col-7">
             <label for="inputQuestion" class="form-label">Question</label>
@@ -27,6 +31,11 @@
     </form>
 
     <a href="/admin/questions/list" type="button" class="btn btn-primary">Back</a>
-    <button type="submit" value="Update" form="edit-question-form" class="btn btn-success">Update</button>
-    <a href="/admin/questions/delete/{{ $question->id }}}" class="btn btn-danger">Delete</a>
+    @if($action === \App\Http\Controllers\Admin\QuestionsBankController::ACTION_EDIT)
+        <button type="submit" value="Update" form="edit-question-form" class="btn btn-success">Update</button>
+        <a href="/admin/questions/delete/{{ $question->id }}}" class="btn btn-danger">Delete</a>
+    @else
+        <button type="submit" value="Create" form="edit-question-form" class="btn btn-success">Create</button>
+    @endif
+
 @endsection
